\docType{data}
\name{data}
\alias{data}
\title{Thickness measurement of silicon waffer}
\format{A data frame with 30 rows and 5 variables:
\describe{
   \item{pos1}{Thickness measurement at Position 1}
}}
\usage{
diamonds
}
\description{
A dataset containing the thickness in micrometer at different positions in a grinder. The variables are as follows:
}
\keyword{datasets}